tests/requirements.txt
black==23.9.1
factory-boy==3.3.0
isort==5.12.0
ruff==0.0.292
pytest==7.4.2
